fix: ci problems for macos and windows
