+++ Cargo.toml
@@ -36,7 +36,6 @@ futures-lite = "2.3.0"
 interprocess = "2.2.3"
 libreauth = "0.17.0"
 log = "0.4.26"
-named-sem = "0.2.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.140"
 serde_yml = "0.0.12"
@@ -57,5 +56,6 @@ log = "0.4.26"
 env_logger = "0.11.6"
 serial_test = "3.2.0"
 
-[target.'cfg(windows)'.dependencies]
-named_pipe = "0.4.1"
+# Platform-specific dependencies for binary (excluded on macOS)
+[target.'cfg(not(target_os = "macos"))'.dependencies]
+named-sem = "0.2.3"
