+++ flake.nix
@@ -127,7 +127,7 @@
             check-case-conflicts.enable = true;
             check-executables-have-shebangs.enable = true;
             check-merge-conflicts.enable = true;
-            check-shebang-scripts-are-executable.enable = true;
+            # check-shebang-scripts-are-executable.enable = true;
             check-toml.enable = true;
             check-yaml.enable = true;
             detect-private-keys.enable = true;
